Move convert function from F0 to B0
The convert function convert special pandas and numpy types
into json.
