+++ src/B0_data_input_json.py
@@ -98,6 +98,49 @@ def convert_special_types(a_dict, prev_key=None):
     return answer
 
 
+def convert(o):
+    """This converts all data stored in dict_values that is not compatible with the json format to a format that is compatible.
+
+    Parameters
+    ----------
+    o :
+        Any type. Object to be converted to json-storable value.
+
+    Returns
+    -------
+    type
+        json-storable value.
+
+    """
+    if isinstance(o, np.int64):
+        answer = int(o)
+    # todo this actually drops the date time index, which could be interesting
+    elif isinstance(o, pd.DatetimeIndex):
+        answer = o.to_frame().to_json(orient="split")
+        answer = "{}{}".format(TYPE_DATETIMEINDEX, answer)
+    elif isinstance(o, pd.Timestamp):
+        answer = str(o)
+        answer = "{}{}".format(TYPE_TIMESTAMP, answer)
+    # todo this also drops the timeindex, which is unfortunate.
+    elif isinstance(o, pd.Series):
+        answer = o.to_json(orient="split")
+        answer = "{}{}".format(TYPE_SERIES, answer)
+    elif isinstance(o, np.ndarray):
+        answer = json.dumps({"array": o.tolist()})
+    elif isinstance(o, pd.DataFrame):
+        answer = o.to_json(orient="split")
+        answer = "{}{}".format(TYPE_DATAFRAME, answer)
+    else:
+        raise TypeError(
+            "An error occurred when converting the simulation data (dict_values) to json, as the type is not recognized: \n"
+            "Type: " + str(type(o)) + " \n "
+            "Value(s): " + str(o) + "\n"
+            "Please edit function CO_data_processing.dataprocessing.store_as_json."
+        )
+
+    return answer
+
+
 def load_json(
     path_input_file, path_input_folder=None, path_output_folder=None, move_copy=False
 ):
